feat: chuyển điểm đạt tối thiểu sang thang thập phân 0–10

+++ src/Services/Exam/Exam.Application/ExamAggregate/Commands/CreateExam/CreateExamCommand.cs
@@ -11,5 +11,5 @@ public record CreateExamCommand(
     Level Level,
     string CategoryId,
     bool IsTimeRestricted,
-    int MinimumPassingScore,
+    decimal MinimumPassingScore,
     string OwnerUserId) : IRequest<ExamDto>;
